desktop: Move VolumeControls into dialogs
